Fixed compilation on NetBSD (and fixed warnings).
git-svn-id: http://svn.aircrack-ng.org/trunk@2757 28c6078b-6c39-48e3-add9-af49d547ecab

+++ src/airodump-ng.c
@@ -4893,7 +4893,7 @@ static int json_get_value_for_name( const char *buffer, const char *name, char *
 		if(*cursor != '\0')
 		{
 			cursor++;
-			while(isspace(*cursor) && *cursor != '\0')
+			while(isspace((int)(*cursor)) && *cursor != '\0')
 			{
 				cursor++;
 			}
@@ -4947,7 +4947,7 @@ static int json_get_value_for_name( const char *buffer, const char *name, char *
 		else
 		{
 			/* Number, supposedly.  Copy as-is. */
-			while(*cursor != ',' && *cursor != '}' && !isspace(*cursor))
+			while(*cursor != ',' && *cursor != '}' && !isspace((int)(*cursor)))
 			{
 				*vcursor = *cursor;
 				cursor++; vcursor++;
